requirements.txt
paramiko
cryptography
pyyaml
colorama
tabulate
